chain.hh -> chain.hh + chain.cc
